backend/pyproject.toml
[project]
name = "knowhive-backend"
version = "0.1.0"
description = "KnowHive FastAPI sidecar backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "aiosqlite>=0.20.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.1",
    "pymupdf>=1.24.0",
    "watchdog>=4.0.0",
    "langchain-ollama>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ragas>=0.2.0",
    "langfuse>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ragas>=0.2.0",
    "langfuse>=2.0.0",
]
